+++ __tests__/clickOutside.test-d.ts
@@ -1,5 +1,5 @@
 import { expectType } from 'tsd';
-import _ from './../index';
+import _ from '../index';
 
 const divElement = document.createElement('div');
 const spanElement = document.createElement('span');
