+++ __tests__/debounce.test-d.ts
@@ -1,4 +1,4 @@
 import { expectType } from 'tsd';
-import _ from './../index';
+import _ from '../index';
 
 expectType<Function>(_.debounce(() => {}, 5000));
