+++ __tests__/pick.test-d.ts
@@ -1,5 +1,5 @@
 import { expectType } from 'tsd';
-import _ from './../index';
+import _ from '../index';
 
 const obj = {
   name: 'al-bur',
