+++ __tests__/throttle.test-d.ts
@@ -1,5 +1,5 @@
 import { expectType } from 'tsd';
-import _ from './../index';
+import _ from '../index';
 
 // NOTE: 보여주기식 테스트 코드...
 expectType<Function>(_.throttle(() => {}, 5000));
